+++ src/Stroke.cpp
@@ -48,7 +48,8 @@ enum class KEY_MODE {
 	S_SCROLL_LEFT = 40,
 	S_SCROLL_RIGHT = 41,
 	S_SCROLL_UP = 42,
-	S_SCROLL_DOWN = 43
+	S_SCROLL_DOWN = 43,
+	S_WINDOW_MINIMIZE = 44
 };
 
 template <int PORTS>
@@ -864,6 +865,12 @@ struct CmdBusboard {
 }; // struct CmdBusboard
 */
 
+struct CmdWindowMinimize : CmdBase {
+	void initialCmd(KEY_MODE keyMode) override {
+		glfwIconifyWindow(APP->window->win);
+	}
+};
+
 // -- GUI --
 
 template<int PORTS>
@@ -995,6 +1002,8 @@ struct KeyContainer : Widget {
 					//if (!cmdBusboard) cmdBusboard = new CmdBusboard;
 					//cmdBusboard->process();
 					break;
+				case KEY_MODE::S_WINDOW_MINIMIZE:
+					processCmd<CmdWindowMinimize>(); break;
 				default:
 					break;
 			}
@@ -1642,6 +1651,7 @@ struct KeyDisplay : StoermelderLedDisplay {
 				//menu->addChild(construct<ModeMenuItem>(&MenuItem::text, "Toggle busboard", &ModeMenuItem::module, module, &ModeMenuItem::idx, idx, &ModeMenuItem::mode, KEY_MODE::S_BUSBOARD));
 				//menu->addChild(construct<ModeMenuItem>(&MenuItem::text, "Toggle engine pause", &ModeMenuItem::module, module, &ModeMenuItem::idx, idx, &ModeMenuItem::mode, KEY_MODE::S_ENGINE_PAUSE));
 				menu->addChild(construct<ModeMenuItem>(&MenuItem::text, "Toggle lock modules", &ModeMenuItem::module, module, &ModeMenuItem::idx, idx, &ModeMenuItem::mode, KEY_MODE::S_MODULE_LOCK));
+				menu->addChild(construct<ModeMenuItem>(&MenuItem::text, "Minimize window", &ModeMenuItem::module, module, &ModeMenuItem::idx, idx, &ModeMenuItem::mode, KEY_MODE::S_WINDOW_MINIMIZE));
 				return menu;
 			}
 		}; // struct SpecialMenuItem
@@ -1764,6 +1774,8 @@ struct KeyDisplay : StoermelderLedDisplay {
 						text = "Scroll down"; break;
 					case KEY_MODE::S_BUSBOARD:
 						text = "Toggle busboard"; break;
+					case KEY_MODE::S_WINDOW_MINIMIZE:
+						text = "Minimize window"; break;
 				}
 
 				Tooltip::step();